+++ src/main/java/com/cheolhyeonpark/step4/domain/Car.java
@@ -1,4 +1,4 @@
-package com.cheolhyeonpark.step4;
+package com.cheolhyeonpark.step4.domain;
 
 import java.util.Objects;
