+++ src/main/java/com/cheolhyeonpark/step4/domain/CarMaker.java
@@ -1,4 +1,4 @@
-package com.cheolhyeonpark.step4;
+package com.cheolhyeonpark.step4.domain;
 
 import org.apache.commons.lang3.StringUtils;
