+++ src/test/java/com/cheolhyeonpark/step4/RacingGameTest.java
@@ -1,5 +1,7 @@
 package com.cheolhyeonpark.step4;
 
+import com.cheolhyeonpark.step4.domain.Car;
+import com.cheolhyeonpark.step4.domain.RacingGame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
