+++ src/test/java/com/cheolhyeonpark/step4/StageTest.java
@@ -1,5 +1,8 @@
 package com.cheolhyeonpark.step4;
 
+import com.cheolhyeonpark.step4.domain.Car;
+import com.cheolhyeonpark.step4.domain.Dice;
+import com.cheolhyeonpark.step4.domain.Stage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
